backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.32.3
httpx==0.28.1
beautifulsoup4==4.12.3
pytest==8.3.2
openai==1.54.0
anthropic==0.39.0
vercel-blob==0.4.2
playwright==1.49.0
google-genai>=1.0.0
